♻️ (feature/mobile-ui) - playlist 관련 중복 로직 hook으로 뺌
